country: [chore] comments & seperation
- Created a new function to improve modularity between functions
